requirements.txt
Flask==3.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
Werkzeug==3.0.1
Jinja2==3.1.2
pathspec==0.11.2
requests==2.31.0
